// backend/src/defaults/enum.ts
export enum RecruitmentType {
  NSS = "nss",
  EMPLOYEE = "employee",
}

export enum RecruitmentStatus {
  CV_REVIEW = "CV Review",
  FIRST_INTERVIEW = "1st Interview",
  ONLINE_EXAMS = "Online Exams",
  HOME_ASSIGNMENT = "Home Assignment",
  TECHNICAL_INTERVIEW = "Technical Interview",
  CONTRACT_TERMS = "Contract Terms",
  CUSTOMER_INTERVIEW = "Customer Interview",
  FACE_TO_FACE_INTERVIEW = "Face to face Interview",
  HIRED = "Hired",
  NOT_HIRED = "Not Hired",
  CONSIDER_FOR_FUTURE = "Consider for Future",
  ON_HOLD = "On Hold",
  IN_QUESTION = "In Question",
  QUIT = "Quit",
  FIRED = "Fired",
}

export enum FailStage {
  CV_REVIEW = "CV Review",
  FIRST_INTERVIEW = "1st Interview",
  ONLINE_EXAMS = "Online Exams",
  HOME_ASSIGNMENT = "Home Assignment",
  TECHNICAL_INTERVIEW = "Technical Interview",
  CONTRACT_TERMS = "Contract Terms",
  CUSTOMER_INTERVIEW = "Customer Interview",
}

export enum RelationshipType {
  MOTHER = "Mother",
  FATHER = "Father",
  SIBLING = "Sibling",
  RELATIVE = "Relative",
  FRIEND = "Friend",
}

export enum EventType {
  BIRTHDAY = "birthday",
  HOLIDAY = "holiday",
  ANNOUNCEMENT = "announcement",
  TRAINING = "training",
  OTHER = "other",
}

export enum PtoStatusType {
  PENDING =  "pending",
  HR_APPROVED = "approved",
  HR_DECLINED = "declined",
  MANAGER_APPROVED = "manager_approved",
  MANAGER_DECLINED = "manager_declined"
}